finish moving to sqlite

+++ src-tauri/migrations/20241028005559_initial_migration.sql
@@ -2,7 +2,7 @@ create table accounts (
   id integer primary key autoincrement,
   name text not null default '',
   account_type text not null default '',
-  balance numeric not null default 0.00,
+  balance integer not null default 0.00,
   opened text not null default (date())
 );
 
@@ -17,7 +17,7 @@ create table transactions (
   category_id integer not null,
   description text not null default '',
   transaction_date text not null default (date()),
-  amount numeric not null default 0.00,
+  amount integer not null default 0.00,
   foreign key (account_id) references accounts (id),
   foreign key (category_id) references categories (id)
 );
